updated type on currency model

+++ src/currency/currency.model.ts
@@ -13,7 +13,6 @@ export class Currency extends Model<Currency, ICurrencyCreationAttributes> {
   @Column({ type: DataType.STRING, allowNull: false })
   name: string // наименование валюты
 
-  @Column({ type: DataType.FLOAT, allowNull: false })
+  @Column({ type: DataType.DOUBLE, allowNull: false })
   rate: number // курс валюты
 }
-
